requirements.txt
altgraph==0.17.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
docopt==0.6.2
et-xmlfile==1.1.0
future==0.18.2
idna==3.4
loguru==0.6.0
numpy==1.23.2
openpyxl==3.0.10
pandas==1.5.0
pefile==2022.5.30
Pillow==9.2.0
pipreqs==0.4.11
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.10
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
pywin32-ctypes==0.2.0
requests==2.28.1
six==1.16.0
ttkbootstrap==1.9.0
urllib3==1.26.12
win32-setctime==1.1.0
xlrd==2.0.1
xlwings==0.27.14
yarg==0.1.9

pdfplumber~=0.7.5
PyYAML~=6.0